admin/requirements.txt
flask=2.3.2
flask-pymongo=2.3.0
pydantic=1.10.9